backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
pandas==2.2.2
openpyxl==3.1.5
python-magic==0.4.27
django-cors-headers==4.4.0
pydantic==2.8.2
python-dotenv==1.0.0
requests==2.31.0
